modif da la migration marks int->float

+++ database/migrations/2017_03_08_000012_create_marks_table.php
@@ -17,7 +17,7 @@ class CreateMarksTable extends Migration
             $table->increments('id');
             $table->integer('student_sport_id')->unsigned();
             $table->foreign('student_sport_id')->references('id')->on('student_sport')->onDelete('cascade');
-            $table->integer('mark');
+            $table->float('mark');
             $table->string('comment')->nullable();
             $table->timestamps();
         });
